refactor(utils): move color-format into color folder
