+++ src/rules/token-border-color.ts
@@ -1,7 +1,7 @@
 import valueParser from 'postcss-value-parser';
 import colorString from 'color-string';
 import { tokenRule } from '../utils/token-rule.js';
-import { detectColorFormat } from '../utils/color-format.js';
+import { detectColorFormat } from '../utils/color/index.js';
 
 export const borderColorRule = tokenRule({
   name: 'design-token/border-color',
